+++ types/element.d.ts
@@ -1,0 +1,10 @@
+import { AnnotationElementOptions } from "./options";
+
+export interface AnnotationElement {
+	x: number,
+	y: number,
+	x2: number,
+	y2: number,
+	options: AnnotationElementOptions
+}
+
